requirements.txt
django==1.9.2
django-angular==0.8.1
djangorestframework==3.3.2
#psycopg2==2.6.1  production db
#lxml==3.5.0    for python parsing
#requests==2.9.1 for python scraping